Refactoring to delete internal package
