+++ stxml/stxml_test.go
